Remove gridstack.js library and modify code accordingly.
Adjust code to no longer depend on the gridstack grid to place elements on the preview pages.
Elements are now placed in absolute pixel values instead of row/column.
Import and Export functions are broken for now since they will be overhauled in a future commit anyway.
Also some minor code refactoring to make it more readable and DRY.