Restructure reporting component to match the new click-workflow. Commented out almost all functionality and GUI-elements.
Three new child-components exist now:
- reportingTemplateSelect
- reportingOverview
- reportingAddIndicator

The old component reportingAddIndicatorModal still exists, but is detached from the rest of the app (as reference, code will be moved to other components piece by piece).
The component reportingModal is now just a wrapper for child components. Code has been moved to reportingOverview.
A lot of Javascirpt code is commented out for now to get to a working state.
Also some dependencies are removed from controllers (to only re-add the ones really needed).

New components are connected by a basic click-workflow as proof-of-concept.
GUI will be rebuild (uncommented) in future commits.