Change implementation for template selection.
Selection is done manually now instead of the <select> and <option> tag.
This might require slightly more code, but is easier since we have the selection inside an ng-repeat (categories).
This is also why ng-options does not work properly (or at least I couldn't get it to work properly).

Also disable template selection button when no template is selected and select template on initialization.