Initial draft for generating template preview directly from scope variables.
For now only one template is implemented (even though the other ones can be selected, too).
Also refine the general settings that are shown on each page and connect them to the preview.

Since the preview is generated with ng-repeat, changes to general settings are directly displayed in the preview.
With this implementation, selecting another template should update the preview automatically without the need of manual adjustments.