Fix bug where echarts instances would display on wrong pages after inserting timestamps.
Reason:
Whenever new pages were needed for inserting, we used the initial version of the template, essentially copying one page multiple times.
Normally this results in an ngRepeat:dupes error, because all copies objects have the same $$hashkey.
The previous workaround was to track pages by index in ngRepeat.

However, this resulted in the bug, that echarts would not move diagrams when new ages were prepended.
The Fix was to use angular.toJSON() and angular.fromJSON() insteas of JSON.stringify() and JSON.parse() and to remove the tracking by index.
The angular-specific functions remove the $$hashkey property so we don't have to use a workaround for duplicate keys anymore.