Prepare diagrams for different timestamps to reduce manual adjustments.
Before, diagrams were prepared once on indicator selection for the most recent timestamp. Then data was changed manually, leading to the problem that the visualMap (legend) didn't fit the data.
To fix this, diagrams are now prepared more often and stored locally.

The map that shows a timeseries is not working properly yet since the change is not calculated during diagram preparation.