Fix a bug where isochrones would not be generated in reporting.
The variable $scope.isUsedInReporting, which indicates if the reachability component is used inside the reporting modal, was set to true when the reporting modal opened. For some reason the function 'kommonitorReachabilityController' is executed again when clicking the button to configure a new reporting poi layer. That sets the variable back to false.
The fix was to set the variable is set to true at a later point, after the mentioned function had been executed.